Configure the project. Do world-simulator is publishing and sensor-control is subscribing.
