+++ base-domain/src/main/java/com/example/basedomain/Point.java
@@ -1,0 +1,14 @@
+package com.example.basedomain;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+public class Point {
+    private int x;
+    private int y;
+
+}
