--- central-unit/settings.gradle.kts
@@ -1,1 +1,0 @@
-rootProject.name = "central-unit"
