--- sensor-control/src/main/java/com/example/sensorcontrol/kafka/Point.java
@@ -1,11 +1,0 @@
-package com.example.sensorcontrol.kafka;
-
-import lombok.*;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Point {
-    private int x;
-    private int y;
-}
