+++ sensor-control/src/main/java/com/example/sensorcontrol/kafka/PositionConsumer.java
@@ -4,6 +4,7 @@ import org.slf4j.LoggerFactory;
 import org.slf4j.Logger;
 import org.springframework.kafka.annotation.KafkaListener;
 import org.springframework.stereotype.Service;
+import com.example.basedomain.Point;
 
 @Service
 public class PositionConsumer {
@@ -19,6 +20,7 @@ public class PositionConsumer {
     {
         try {
             LOGGER.info("listener triggered");
+            LOGGER.info(point.toString());
 
         } catch (RuntimeException e) {
             throw new RuntimeException(e);
