+++ sensor-control/src/main/resources/application.properties
@@ -7,6 +7,6 @@ spring.kafka.consumer.auto-offset-reset=latest
 spring.kafka.consumer.key-deserializer=org.apache.kafka.common.serialization.StringDeserializer
 spring.kafka.consumer.value-deserializer=org.springframework.kafka.support.serializer.JsonDeserializer
 spring.kafka.properties.spring.json.trusted.packages=*
-spring.kafka.topic.consume.name=TargetPointPosition2
+spring.kafka.topic.consume.name=TargetPointPosition3
 spring.kafka.topic.produce.position.name=TowerPosition
 spring.kafka.topic.produce.position.info.name=TargetBearingPosition
