+++ settings.gradle.kts
@@ -1,1 +1,5 @@
 rootProject.name = "target_location_estimation"
+include("world-simulator")
+include("base-domain")
+include("central-unit")
+include("sensor-control")
