--- world-simulator/src/main/java/com/example/worldsimulator/data/Point.java
@@ -1,12 +1,0 @@
-package com.example.worldsimulator.data;
-
-
-import lombok.*;
-
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-public class Point {
-    private int x;
-    private int y;
-}
