+++ world-simulator/src/main/resources/application.properties
@@ -3,7 +3,7 @@ server.port=8083
 spring.kafka.bootstrap-servers= localhost:9092
 spring.kafka.producer.key-serializer= org.apache.kafka.common.serialization.StringSerializer
 spring.kafka.producer.value-serializer= org.springframework.kafka.support.serializer.JsonSerializer
-spring.kafka.topic.name=TargetPointPosition2
+spring.kafka.topic.name=TargetPointPosition3
 target.heading=360
 target.initial.position.X = -1
 target.initial.position.Y = 5
